Update stage-backend-v2-ebs branch.

+++ emogo/apps/stream/serializers.py
@@ -22,7 +22,6 @@ from emogo.apps.notification.views import NotificationAPI
 from emogo.apps.notification.models import Notification
 from django.db import IntegrityError
 from django.core.exceptions import ObjectDoesNotExist
-
 from asgiref.sync import async_to_sync
 from channels.layers import get_channel_layer
 from functools import reduce
@@ -71,7 +70,6 @@ def delete_comments_and_broadcast(
         remove_comments_notifications(comment_ids, notification_type)
         broadcast_by_type(data, "private", stream.id)
         broadcast_by_type(data, "public", stream.id)
-
 
 
 class StreamSerializer(DynamicFieldsModelSerializer):
@@ -129,6 +127,12 @@ class StreamSerializer(DynamicFieldsModelSerializer):
                         'type': {'required': True, 'allow_blank': False, 'allow_null': False}
                         # 'image': {'required': True, 'allow_blank': False, 'allow_null': False}
                         }
+
+    # def validate_folder(self, value):
+    #     if value and value.owner != self.context["request"].user:
+    #         raise serializers.ValidationError(messages.MSG_FOLDER_OWNER_NOT_VALID)
+    #     else:
+    #         return value
 
     def validate(self, attrs):
         # This code is run only in case of update through the PATCH method:
@@ -218,10 +222,10 @@ class StreamSerializer(DynamicFieldsModelSerializer):
                 self.create_content(self.instance)
         #3  Update the status of  all collaborator is Inactive When Stream is Global otherwise Collaborator Status is Active
         if self.context['version']:
-            collaborator_list = self.instance.collaborator_list.exclude(status='Unverified')    
+            collaborator_list = self.instance.collaborator_list.exclude(status='Unverified')
         else:
-            collaborator_list = self.instance.collaborator_list.all()    
-        if collaborator_list.__len__ > 0:
+            collaborator_list = self.instance.collaborator_list.all()
+        if collaborator_list.__len__() > 0:
             stream_type = self.validated_data.get('type')
             if stream_type == 'Public':
                 # When Stream is (Public -> Global) and (Private -> Global), (Global -> Public) 
@@ -289,12 +293,12 @@ class StreamSerializer(DynamicFieldsModelSerializer):
 
         collaborator_list = self.initial_data.get('collaborator')
         self.owner_collaborator(stream, collaborator_list)
-        collaborators = map(self.save_collaborator, collaborator_list,
-                            itertools.repeat(stream, collaborator_list.__len__()))
-        
+        collaborators = list(map(self.save_collaborator, collaborator_list,
+                            itertools.repeat(stream, collaborator_list.__len__())))
         if stream.collaborator_list.count() == 1:
-            if  stream.collaborator_list.all()[0].created_by == self.context.get('request').user and \
-                stream.collaborator_list.all()[0].phone_number == self.context.get('request').user.username:
+            if stream.collaborator_list.all()[0].created_by == self.context.get('request').user and \
+                stream.collaborator_list.all()[0].phone_number == self.context.get('request').user.username and \
+                    self.instance:
                 self.instance.collaborator_list.filter().delete()
         else:
             return collaborators
@@ -340,7 +344,7 @@ class StreamSerializer(DynamicFieldsModelSerializer):
         :return: Add Stream content
         """
         content_list = self.initial_data.get('content')
-        contents = map(self.save_content, content_list, itertools.repeat(stream, content_list.__len__()))
+        contents = list(map(self.save_content, content_list, itertools.repeat(stream, content_list.__len__())))
         return contents
 
     def save_content(self, data, stream):
@@ -377,7 +381,8 @@ class StreamSerializer(DynamicFieldsModelSerializer):
             created_by=self.context.get('request').user,
             height=self.validated_data.get('height', 300),
             width=self.validated_data.get('width', 300),
-            color = self.validated_data.get('color')
+            color=self.validated_data.get('color'),
+            # folder=self.validated_data.get("folder", None)
         )
         # stream.save()
         # Update any_one_can_edit flag is type is Public
@@ -481,7 +486,7 @@ class ViewStreamSerializer(StreamSerializer):
                         setattr(instance, 'user_id', user.get('id'))
                         setattr(instance, 'user_image', user.get('user_data__user_image'))
                     # If some collaborator are not registered.
-                    elif not user.get('username').endswith(instance.phone_number) and not instance.phone_number in map(lambda x: x.phone_number, list_of_instances):
+                    elif not user.get('username').endswith(instance.phone_number) and not instance.phone_number in [x.phone_number for x in list_of_instances]:
                         setattr(instance, 'name', instance.name)
                         setattr(instance, 'user_profile_id', None)
                         setattr(instance, 'user_id', None)
@@ -560,7 +565,7 @@ class ViewStreamSerializer(StreamSerializer):
         # Find the logged in user and fetch current user's followers 
         user_id = self.context.get('request').user.id
         try:
-            return [{'id': x.user.id, 'user_profile_id': x.user.user_data.id, 'user_image': x.user.user_data.user_image,'full_name': x.user.user_data.full_name, 'display_name': x.user.user_data.display_name, 'is_following': True if user_id in  map(lambda y: y.follower.id, x.user.user_liked_followers) else False } for x in obj.total_like_dislike_data ]
+            return [{'id': x.user.id, 'user_profile_id': x.user.user_data.id, 'user_image': x.user.user_data.user_image,'full_name': x.user.user_data.full_name, 'display_name': x.user.user_data.display_name, 'is_following': True if user_id in  [y.follower.id for y in x.user.user_liked_followers] else False } for x in obj.total_like_dislike_data ]
         except AttributeError:
             return None
 
@@ -918,8 +923,8 @@ class MoveContentToStreamSerializer(ContentSerializer):
         self.initial_data['contents'].update(upd=datetime.datetime.now())
         for stream in self.initial_data.get('streams'):
             self.initial_data['thread'] = random_generator()
-            map(self.add_content_to_stream, self.initial_data.get('contents'),
-                                itertools.repeat(stream, self.initial_data.get('contents').__len__()))
+            list(map(self.add_content_to_stream, self.initial_data.get('contents'),
+                                itertools.repeat(stream, self.initial_data.get('contents').__len__())))
 
             if self.context['version']:
                 collab_list = stream.collaborator_list.filter(status= 'Active')
@@ -944,7 +949,7 @@ class MoveContentToStreamSerializer(ContentSerializer):
         :return: Function add content to stream
         """
         # Create Stream and content
-        obj , created = StreamContent.objects.get_or_create(content=content, stream=stream, user=self.context.get('request').user)
+        obj, created = StreamContent.objects.get_or_create(content=content, stream=stream, user=self.context.get('request').user)
         # Set True in have_some_update field, When user move content to stream
         obj.thread = self.initial_data.get("thread")
         obj.save()
@@ -1404,6 +1409,7 @@ class SeenIndexSerializer(DynamicFieldsModelSerializer):
 def set_have_some_update_true(stream):
     NewEmogoViewStatusOnly.objects.filter(stream=stream).update(have_some_update = True)
     return True
+
 
 class FolderSerializer(DynamicFieldsModelSerializer):
     """
